requirements.txt
# requirements.txt
pandas>=2.0.0
boto3>=1.28.0
datasets>=2.14.0
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-core>=0.1.0
ragas>=0.1.0
deepeval>=0.20.0
nltk>=3.8.0
rouge-score>=0.1.2
python-Levenshtein>=0.21.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
openpyxl>=3.1.0